Reverted Shattrath changes, apparently they are now accessible even without picking a faction. It will still hide buckets where you are below Neutral with either Aldor or Scryers.
Trying to fix TomTom arrow, but it seems to be a but faulty as it is right now. Relogging usually makes the arrow appear after clicking on the nodes.

Fixed an issue where browsing the calendar could trigger loading the other modules for Lunar or Midsummer. This shouldn't occur any longer.